Updated utils
